requirements.txt
ipykernel==6.29.5
datasets==2.16.1
openai==1.30.1
PyPDF2==3.0.1
transformers==4.38.0
langchain-experimental==0.0.64
langchain-openai==0.1.10
python-dotenv==1.0.1
PyYAML==6.0.1
coloredlogs==15.0.1
mdc==1.2.1
pytest==8.1.2
Wand==0.6.13
strictyaml==1.7.3

promptflow-core==1.14.0
promptflow-devkit==1.14.0
promptflow-tracing==1.14.0
promptflow-azure==1.14.0
promptflow-evals==0.3.1

# Fine-tuning
azure-ai-ml==1.19.0
azure-identity==1.17.1
mlflow==2.15.1
azureml-mlflow==1.57.0
matplotlib==3.9.1
numpy==1.26.4

# To run notebooks from the command line
papermill==2.6.0
black[jupyter]==24.8.0

# Azure ML Infra
azureml-core
msrestazure
azure-mgmt-resource

# Infra tests
pytest==8.1.2
requests==2.32.3
